% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate.R
\name{bind_tetra}
\alias{bind_tetra}
\title{Bind a list of tetra matrices to one matrix (n x 136)}
\usage{
bind_tetra(x)
}
\arguments{
\item{x}{a list of tetramer matrices}
}
\value{
a matrix with uniquely named rows
}
\description{
Bind a list of tetra matrices to one matrix (n x 136)
}
