% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast.R
\name{blastHitText}
\alias{blastHitText}
\title{Trim a blast hit description string to a particular length with it's associated
hit score and accession name.}
\usage{
blastHitText(x, len = 30, hsp_bit_score_min = 75,
  noHitText = "no significant hits")
}
\arguments{
\item{x}{blast tibble}

\item{len}{numeric, the maximum output string length}

\item{hsp_bit_score_min}{numeric, hits below this are flagged with noHitText}

\item{noHitText}{the value to return for failed hits}
}
\value{
the tibble with an added column 'hit_tetxt'
}
\description{
Trim a blast hit description string to a particular length with it's associated
hit score and accession name.
}
