% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{cexfun}
\alias{cexfun}
\title{Returns a numeric vector suitable of cex values for plot symbols}
\usage{
cexfun(x = 1:10, big = 1.5, small = 0.5, what = c("simple", "bigfirst",
  "linear")[1])
}
\arguments{
\item{x}{a numeric vector - we only care about its length}

\item{big}{numeric, cex values for big symbols}

\item{small}{numeric, cex values for small symbols}

\item{what}{character
\itemize{
  \item{'simple' are small size}
  \item{'linear' points are scales from big to small}
  \item{'bigfirst' first point is big all others small}
}}
}
\value{
numeric vector suitable for cex
}
\description{
Returns a numeric vector suitable of cex values for plot symbols
}
