% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{clean_sequence}
\alias{clean_sequence}
\title{Tidy an input FASTA string}
\usage{
clean_sequence(x, dict = Biostrings::DNA_ALPHABET, replacement = "N")
}
\arguments{
\item{x}{the input contig sequence OR a simple string}

\item{dict}{character, the dictionary of allowed characters}

\item{replacement}{character, the replacement value for disallowed characters}
}
\value{
either a contig squence or a character string - matches the input
}
\description{
Tidy an input FASTA string
}
