% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{getBlastWindow}
\alias{getBlastWindow}
\title{Retrieve the window specification from a window name.}
\usage{
getBlastWindow(X, windowName = "unknown_1-5000", delim = "-")
}
\arguments{
\item{X}{a TetramerPCA object}

\item{windowName}{the full name of the window ("contigName_start-end")}

\item{delim}{the delimiter between the start and end of the window}
}
\value{
the sequence within the window
}
\description{
Retrieve the window specification from a window name.
}
