% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{getExplanation}
\alias{getExplanation}
\title{Retrieve the normalized 'explanation' of variance for each PC as eigenvalues}
\usage{
getExplanation(X, n = length(X$pick), asVariance = TRUE)
}
\arguments{
\item{X}{the Tetramers R6 class object}

\item{n}{the number of elements to return, by default the number of PCs}

\item{asVariance}{if TRUE then use the variance (which is the eigenvalue) instead of the}
}
\value{
a vector of eigenvalues
}
\description{
Returns the 'explanations' of variance as derived from the sdev component of the
 PC (principal components)
}
