% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{getTetramerRange}
\alias{getTetramerRange}
\title{Compute pretty [xy] ranges for TetramerPCA data}
\usage{
getTetramerRange(X, n = length(X$pick))
}
\arguments{
\item{X}{A TetramerPCA class object}

\item{n}{the number of ranges to return, by default the length of PCs we pick}
}
\value{
a 2 column by n row matrix of min and max ranges
}
\description{
Compute pretty [xy] ranges for TetramerPCA data
}
