% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{input_name}
\alias{input_name}
\title{Extract a name from a filename}
\usage{
input_name(filename = c("foo.fasta", "bar.fasta.gz"))
}
\arguments{
\item{filename}{character, one or more filenames}
}
\value{
one or more names
}
\description{
Extract a name from a filename
}
