% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{makeCombinations}
\alias{makeCombinations}
\title{Create a  vector of the default combinations of A, C, G and T}
\usage{
makeCombinations(what = c("tetramers", "tetramers-all")[1], sorted = TRUE,
  as.DNAStringSet = FALSE, lowercase = FALSE)
}
\arguments{
\item{what}{character, the name of the combinations, "tetramers" means A,C,G, and T
while "tetramers-all" adds "N" to the tetramers list.}

\item{sorted}{logical, sort into lexical order (AAAA, AAAC, ..., TTTT)}

\item{as.DNAStringSet}{logical, if TRUE then return a DNAStringSet otherwise a vector}

\item{lowercase}{logical, if TRUE then use lower case letters}
}
\value{
a character vector or DNStringSet
}
\description{
Create a  vector of the default combinations of A, C, G and T
}
