% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate.R
\name{normalize_tetra}
\alias{normalize_tetra}
\title{Normalize the tetra sequences}
\usage{
normalize_tetra(x)
}
\arguments{
\item{x}{a list of matrices or a single matrix}
}
\value{
a list or matrix of the same form as the input
}
\description{
This function normalizes the tetramer counts by window locations
You can visualize each window location by one row of the census matrix
Normalization occurs across each row, with the max count scoring the highest
value toward 1.0 and the least represented as toward 0.0.
If there are no hits of the any TETRAMERS (all Ns in the original)
then the normalized score for each tetramer in the row is 0.
}
