% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate.R
\name{pca_tetra}
\alias{pca_tetra}
\title{Perform Principal Component Analysis on the normalized, reduced, tetramer matrix}
\usage{
pca_tetra(x, scale. = TRUE, ...)
}
\arguments{
\item{x}{numeric, the (n x 136) matrix of normalized counts}

\item{scale.}{see \code{\link{prcomp}}}

\item{...}{further arguments for \code{\link{prcomp}}}
}
\description{
Perform Principal Component Analysis on the normalized, reduced, tetramer matrix
}
