% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{pchfun}
\alias{pchfun}
\title{Returns a numeric vector suitable of pch values for plot symbols}
\usage{
pchfun(x = 1:10, what = 24)
}
\arguments{
\item{x}{a numeric vector - we only care about its length}

\item{what}{the symbol identifier}
}
\value{
the value for what repeated
}
\description{
Returns a numeric vector suitable of pch values for plot symbols
}
