% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_PCvPC}
\alias{plot_PCvPC}
\title{Plot one PCx vs PCy pairing}
\usage{
plot_PCvPC(X, PC = c("PC1", "PC2"), MAXSTRING = 35,
  noHitText = "no significant hits", addBlast = !is.null(X$blast_data))
}
\arguments{
\item{X}{a TetramerPCA class object}

\item{PC}{two element vector of PC names}

\item{MAXSTRING}{numeric, the maximum displayed length of a contig name}

\item{noHitText}{the value to return for failed hits}

\item{addBlast, }{logical}
}
\description{
Plot one PCx vs PCy pairing
}
