% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_tetramers}
\alias{plot_tetramers}
\title{Plot the Tetramers results with our without blast annonations}
\usage{
plot_tetramers(X, filename = file.path(X$out_dir, paste0(paste(X$name,
  X$parameters[["WINDOW"]], X$parameters[["STEP"]], "PC", sep = "-"), ".pdf")),
  dim = c(width = 8, height = 8), MAXSTRING = 35,
  addBlast = !is.null(X$blast_data), ...)
}
\arguments{
\item{X}{a Tetramers class object}

\item{filename}{character, the name of the output file}

\item{dim}{a 2 element vector of output width and height in inches}

\item{MAXSTRING}{numeric, the maximum displayed length of a contig name}

\item{addBlast}{logical for suppressing blast annotations}

\item{...}{further arguments for the \code{pdf} device}
}
\value{
named logical indicating that the ouptut file exists
}
\description{
Plot the Tetramers results with our without blast annonations
}
