% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_blast_output}
\alias{read_blast_output}
\title{Read a blast xml output file}
\usage{
read_blast_output(filename)
}
\arguments{
\item{filename}{character, the name of the xml file to read}
}
\value{
a lit with two elements: info and data either of which may be NULL
}
\description{
Read a blast xml output file
}
