% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_fastafile}
\alias{read_fastafile}
\title{Read a FASTA file and possibly tidy it}
\usage{
read_fastafile(X, clean = TRUE)
}
\arguments{
\item{X}{Tetramers-class object}

\item{clean}{logical, if TRUE then tidy the input}
}
\value{
X Tetramers-class object
}
\description{
Read a FASTA file and possibly tidy it
}
