% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate.R
\name{reduce_tetra}
\alias{reduce_tetra}
\title{Reduce the tetra census table to include just the 136 complements (for tetramers)
by pooling reverse complements.}
\usage{
reduce_tetra(x, comp = TETRA[["RCTETRAMERS"]], keep = TETRA[["TETRAMERS"]])
}
\arguments{
\item{x}{a list of matrices or a single matrix (nWindow x 256)}

\item{comp}{a named character vector of reverse complements}

\item{keep}{a character vector of name tetramers to retain}
}
\value{
a list or matrix with just 136 columns
}
\description{
Reduce the tetra census table to include just the 136 complements (for tetramers)
by pooling reverse complements.
}
