% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast.R
\name{run_blast_app}
\alias{run_blast_app}
\title{Run blast and extract the results}
\usage{
run_blast_app(X, verbose = FALSE)
}
\arguments{
\item{X}{a Tetramers R6 class object}

\item{verbose}{logical, if TRUE output messages for debugging purposes}
}
\value{
the updated input object, blast results stored in the 'blast_data' field
 and blast info stored in 'blast_info' field
}
\description{
Run blast and extract the results
}
