% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{select_outliers}
\alias{select_outliers}
\title{Selects the outliers for PCs 1-nPC using a tidy-approach}
\usage{
select_outliers(X)
}
\arguments{
\item{X}{a Tetramers R6 class reference}
}
\value{
the input x
}
\description{
It's a select-and-remove process.  First order by PCx. Select the outliers and
remove contigs with matching names.  Then select a second set of outliers.  So we now have
2-low (negative) and 2-high (positive) outliers along axis PCx.  With this winnowed
data set order by PCy. Select the outliers and remove contigs with matching names.  Then 
select a second set of outliers along this axis.  Now we have the same kind of double set.
}
