% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate.R
\name{tabulate_windows}
\alias{tabulate_windows}
\title{Tabulate the tetramers within a window advanced by step}
\usage{
tabulate_windows(x, tWindow = 1600, tStep = 200, tWidth = 4)
}
\arguments{
\item{x}{a DNAStringSet (or DNAString)}

\item{tWindow}{numeric, the width of the sliding window in base units}

\item{tStep}{numeric, the number of bases to advance between successive window locations}

\item{tWidth}{numeric, the width of the genome unit, for tetramer analysis it's 4}
}
\value{
for each DNAString a list of \code{status} (0/1 where 0 = ok) and \code{x} a
   tabulation matrix.  In the matrix each row corresponds to a window location along
   the length of the sequence.
}
\description{
This generally is called with a DNAStringSet, and then is called recursively
with DNAString.
}
