% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tetramers-package.R
\docType{package}
\name{tetramers-package}
\alias{tetramers-package}
\title{tetramers}
\description{
tetramers detects outlier regions in genomes in terms of tetramer composition,
which are indicative of either contamination or gene transfer. Tetramers along
a contig are tabulated by sliding a window along the length each contig.
}
