% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tetramers-class.R
\docType{data}
\name{Tetramers}
\alias{Tetramers}
\title{A base container class for tetramer analysis}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Tetramers
}
\arguments{
\item{x}{character, filename for FASTA format sequences}

\item{name}{character, by default extracted from filename}

\item{output_dir}{the name of the output directory, if not provided then it is automatically a
subdirectory of the input file's directory}

\item{parameters}{numeric, named vector of sleection parameters
\itemize{
  \item{window window size in bases, default = 1600}
  \item{window step the distance in bases to advance the window,
    default = 200}
  \item{width sort of dumb, but tetramers are in 4-bases, default = 4}
  \item{hsp_bit_score_min hits below this are flagged with noHitText,
    default = 75}
}}

\item{pick}{numeric, named numeric vector of PC id with number to pick.
By default 2 extremes  (2 pos, 2 neg) for PC1-8.}

\item{select_method}{character, by default "tidy". Ignored.}

\item{blast_options}{character, the blast options}

\item{...}{further arguments for \code{read_fastafile}}

\item{verbose}{logical, if TRUE output some chatter}

\item{form}{character, either 'matrix' or 'table' (default)}

\item{npc}{numeric, the first npc components are returned}

\item{filename}{character, xml filename}

\item{verbose}{logical, if TRUE output messages for debugging purposes}
}
\value{
matrix or table of normalized tetramer counts

matrix or table of principle components

matrix or table of principle component roational loadings

a table of [cname, wname, PC, pick, sequence] which is an
  augmented version of the outlers table

logical, TRUE if successful
}
\description{
R6 base class to hold all of the goodies

Create a new Tetramers container object

Read a FASTA file

tabulate tetramers, compute PCA and select outliers

Retrieve the normalized tabulations as a matrix or table

Retrieve the principal components as a matrix or table

Retrieve the principal component rotational loadings as a matrix or table

Retrieve outlier sequences

load blast results

Run blast on the outlier fasta

Run blast on the outlier fasta or simply load existsing blast ouptut

Run blast on the outlier fasta
}
\section{Fields}{

\describe{
\item{\code{filename}}{character, the fully qualified FASTA filename}

\item{\code{name}}{character, parsed from the filename by default but settable}

\item{\code{out_dir}}{character, the fully qualified output path, possibly created}

\item{\code{parameters}}{named numeric vector, composed of 
\itemize{
  \item{window window size in bases, default = 1600}
  \item{window step the distance in bases to advance the window,
    default = 200}
  \item{width sort of dumb, but tetramers are in 4-bases, default = 4}
  \item{hsp_bit_score_min hits below this are flagged with noHitText,
    default = 75}
}}

\item{\code{pick}}{numeric, named vector of number out outliers to pick per PC - generally 2.}

\item{\code{blastOpts}}{list, a named list of blast options}

\item{\code{blast}}{any, blast results}

\item{\code{blastinfo}}{any, info on the blast application}

\item{\code{rawdata}}{any, raw counts}

\item{\code{data}}{matrix,}

\item{\code{fail}}{list, list of failed contigs (generally too short)}

\item{\code{PC}}{list, prcomp results of which we are interested in "x"}

\item{\code{outliers}}{tibble, selected outliers per PC}
}}

\keyword{datasets}
