% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast.R
\name{trim_QueryDef}
\alias{trim_QueryDef}
\title{Trim a blast query name to a reasonable length}
\usage{
trim_QueryDef(x, len = 30)
}
\arguments{
\item{x}{blast tibble}

\item{len}{numeric, the maximum string length allowed}
}
\value{
a character vector
}
\description{
Trim a blast query name to a reasonable length
}
