% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{trimLegendText}
\alias{trimLegendText}
\title{trims input text vector to length and adds etc if needed in such a way that
\code{nchar(paste(newX, etc)) < length}}
\usage{
trimLegendText(x, length = 35, etc = "...", strip.white = TRUE)
}
\arguments{
\item{x}{character vector of one or more strings}

\item{length}{numeric, maximum string length allowed}

\item{etc}{character, the flag to indicate removed characters (a suffix)}

\item{strip.white}{logical, if TRUE strip newlines and tabs}
}
\value{
input chacater vector with elements possibly abbreviated
}
\description{
trims input text vector to length and adds etc if needed in such a way that
\code{nchar(paste(newX, etc)) < length}
}
