% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{write_tetra}
\alias{write_tetra}
\title{Write all of the known results}
\usage{
write_tetra(X, path = X$out_dir, what = c("input", "info", "count", "pca",
  "fail", "loading", "outliers"), verbose = FALSE)
}
\arguments{
\item{X}{Tetramers-class object}

\item{path}{the output path - created if it doesn't exist}

\item{what}{character - a list of items to write
\itemize{
 \item{input the input FASTA file if it is available and doesn't already exist in path}
 \item{info a dump of input parameters to a yaml}
 \item{count tetramer counts}
 \item{pca pca values}
 \item{fail listing of failed tetramer windows}
 \item{loading pca loading values}
 \item{outliers table of outliers}
}}

\item{verbose}{logical, if TRUE then output messages}
}
\description{
Write all of the known results
}
